[libc++] Split __shared_count out of <__memory/shared_ptr.h>
